Split up Grunt Tasks
Breaking up these files makes it easier to maintain, debug, change and
learn.

Note: This hasn't been tested successfully so some changes may be
needed.
