+++ package.json
@@ -43,6 +43,7 @@
   "devDependencies": {
     "connect-livereload": "~0.4.0",
     "grunt": "~0.4.4",
+    "load-grunt-tasks": "^0.4.0",
     "grunt-angular-templates": "^0.5.4",
     "grunt-asset-injector": "^0.1.0",
     "grunt-autoprefixer": "~0.7.2",
